Move composition to separate file
