+++ test/Decapodes2/runtests.jl
@@ -1,0 +1,6 @@
+using Test
+
+@testset "Composition" begin
+  include("composition.jl")
+end
+
